requirements.txt
asgiref==3.8.1
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
dj-database-url==2.3.0
Django==5.2
google-auth==2.40.2
google-auth-oauthlib==1.2.2
gunicorn==23.0.0
idna==3.10
oauthlib==3.2.2
packaging==25.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
python-decouple==3.8
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
whitenoise==6.9.0
paypalrestsdk==1.14.0
python-dotenv==1.0.1